Refactor navbar
1. context menu 파일 분할
2. friend list 파일 분할
